[ALTI Calibs] Adaptations for ALTI based calibrations
The orchestrator is driving the calibration now. At the beginning, it
asks the calib RC instances to publish information about the chosen
`CalibAlg` and the number of iterations. The orchestrator then
requests that the `CalibAlg` resets the interation counter, and then
in a loop asks the calib RC instances to configure, acquire and
unconfigure and performs the ALTI interactions in between.

The defined ALTI interactions go into a `nsw::commands::Commands`
struct that contains three vectors of commands:
`before` (executed before the `configure` step), `during` (executed
after `configure` and before `acquire`), and `after` (executed after
`acquire` and before `unconfigure`)

Notes:
* Start counter at 0 not  -1
* Total should be at least 1 for a single iteration now

If the `STOP` transition is received before the calibration loop
terminates, the Orchestrator will halt subsequent iterations, but at
the moment it does not request that the `CalibAlg` interrupt its
current iteration.
For any long-running calibration, this will block `STOP` from
completing until the current iteration terminates.
To bypass this, one may issue the `SHUTDOWN` command.