Use device manager from NSWConfig
This discussion was triggered by the issue opening too many OPC connections. It was decided that the calibration application will obtain a reference to the device manager from the the NSWConfig object. Eventually all calibration should use the hardware interfaces provided that way. This commit adapts all existing hardware interfaces to take them from the device manager instead. It also passes the device manager to the CalibAlg parent class so all calibrations can use it in the future. This change allows to move parts of the initialization from the setup function to the constructor as well.